fix ConcurrentModificationExcception. Here are some observations and explainations:
1. ConcurrentModificationExcception indicates that someone removes some
elements from a set or a map while he's doing iteration on this set or
map;
2. When I disable the removal in `invoke` method of `messageDataBuffer`
in class `Socket`, I found my test ran without exceptions.

So I'm sure something goes wrong with this `messageDataBuffer`. On the
other hand, the ConcurrentModificationExcception shows that something is
wrong when calculating hash code of an abstract set. And in fact, this
abstract set is `messageDataBuffer`. Looking into what `hashCode()`
method does, I found that this method iterates every elements of the
hash set.

And I got the conclusion: while a thread is calculating hashCode of
`messageDataBuffer` and iterating its elements here, there's another
thread which is trying to remove an element from this hash set.

Soon I made a solution: to subclass HashSet and overload two methods
`hashCode()` and `remove`. The overloading is just to lock the hash set.
And my test runs very well.

Since the problem is because of hashCode and remove, so I made an
improved solution, that is to change the hash code of hash set at the
time when a new element is added or a existed element is removed.